CG, size_t add

+++ gulkov.mikhail/I4/CountPositiveCols.cpp
@@ -1,6 +1,6 @@
 #include "CountPositiveCols.hpp"
 
-unsigned int countPositiveCols(const int *array, size_t cols, size_t rows)
+size_t countPositiveCols(const int *array, size_t cols, size_t rows)
 {
   unsigned int positive_cols = 0;
   for (size_t i = 0; i < cols; i++)
